requirements.txt
# Multi-Tool Orchestrator Agent Requirements
# ==========================================

# FastMCP for modern MCP protocol handling
fastmcp>=0.9.0

# OpenAI for LLM capabilities
openai>=1.3.0

# Async HTTP client for API calls
aiohttp>=3.8.0

# Environment variable management
python-dotenv>=1.0.0

# JSON schema validation
jsonschema>=4.19.0

# Data validation and settings
pydantic>=2.5.0

# Logging utilities
structlog>=23.2.0

# UUID generation for session tracking
uuid

# Additional utilities
typing-extensions>=4.8.0

# Optional: For enhanced logging and monitoring
rich>=13.0.0